+++ pkg/vm/engine/disttae/transfer.go
@@ -35,6 +35,7 @@ import (
 	v2 "github.com/matrixorigin/matrixone/pkg/util/metric/v2"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/disttae/logtailreplay"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/engine_util"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/common"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/mergesort"
 )
@@ -348,7 +349,7 @@ func doTransferRowids(
 	expr := ConstructInExpr(ctx, pkColumName, searchPKColumn)
 
 	var blockList objectio.BlockInfoSlice
-	if _, err = TryFastFilterBlocks(
+	if _, err = engine_util.TryFastFilterBlocks(
 		ctx,
 		table.db.op.SnapshotTS(),
 		table.GetTableDef(ctx),
